requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
libgravatar==1.0.4
limits==3.14.1
Mako==1.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
unicorn==2.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
